app-web/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
Werkzeug==2.3.7
scikit-learn==1.3.0
scipy==1.11.3
matplotlib==3.7.2
joblib==1.3.2
imutils==0.5.4
scikit-image==0.21.0